Added ability to work with multiple sessions on the same account

+++ src/main/java/com/azino/project/service/impl/SignOutServiceImpl.java
@@ -3,9 +3,12 @@ package com.azino.project.service.impl;
 import com.azino.project.model.ShoppingBasket;
 import com.azino.project.service.ShoppingBasketService;
 import com.azino.project.service.SignOutService;
+import com.azino.project.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import javax.servlet.http.HttpSession;
 
 @Service
 @Transactional
@@ -14,11 +17,19 @@ public class SignOutServiceImpl implements SignOutService {
     @Autowired
     private ShoppingBasketService shoppingBasketService;
 
+    @Autowired
+    private UserService userService;
+
     @Override
-    public void signOut(String userName) {
+    public void signOut(String userName, HttpSession session) {
         ShoppingBasket shoppingBasket = shoppingBasketService.findByUserName(userName);
         if(shoppingBasket != null) {
-            shoppingBasketService.deleteAllItemsFromShoppingBasket(shoppingBasket.getId());
+            userService.deleteSessionFromActiveSessions(userName, session.getId());
+            if(userService.getAllActiveSessions(userName).isEmpty()) {
+                shoppingBasketService.deleteAllItemsFromShoppingBasket(shoppingBasket.getId());
+                shoppingBasketService.delete(shoppingBasket);
+            }
+            session.invalidate();
         }
     }
